Moved topBar to layout.tsx
